backend/requirements.txt
Django==5.1.8
djangorestframework==3.15.2
drf-spectacular==0.28.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-decouple==3.8
